+++ ft_strjoin.c
@@ -14,10 +14,10 @@
 
 char	*ft_strjoin(char const *s1, char const *s2)
 {
-	unsigned int	len1;
-	unsigned int	len2;
-	unsigned int	len;
-	char			*str;
+	size_t	len1;
+	size_t	len2;
+	size_t	len;
+	char	*str;
 
 	len1 = ft_strlen(s1);
 	len2 = ft_strlen(s2);
